requirements.txt
django==5.1.3
asgiref==3.8.1
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow==11.0.0
pytest==8.3.4
pytest-django==4.11.1
pytest-cov==6.0.0
django-storages==1.14.4
boto3==1.35.95
supabase==2.11.1

# Production dependencies
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
